+++ icevision/visualize/__init__.py
@@ -1,4 +1,8 @@
 from icevision.visualize.utils import *
 from icevision.visualize.draw_data import *
 from icevision.visualize.show_data import *
-from icevision.visualize.wandb_img import *
+
+from icevision.soft_dependencies import SoftDependencies
+
+if SoftDependencies.wandb:
+    from icevision.visualize.wandb_img import *